requirements.txt
requests
django
pytest
pytest-django
djangorestframework
djangorestframework-simplejwt

#personal customization
flake8
black
isort
